# Conflicts:
#	src/main/java/org/ever/_4ever_be_business/config/DataPreloader.java

+++ src/main/java/org/ever/_4ever_be_business/sd/dto/response/InventoryCheckResultItemDto.java
@@ -16,13 +16,13 @@ public class InventoryCheckResultItemDto {
     private String itemName;
 
     @JsonProperty("requiredQuantity")
-    private Long requiredQuantity;
+    private Integer requiredQuantity;
 
     @JsonProperty("inventoryQuantity")
-    private Long inventoryQuantity;
+    private Integer inventoryQuantity;
 
     @JsonProperty("shortageQuantity")
-    private Long shortageQuantity;
+    private Integer shortageQuantity;
 
     @JsonProperty("statusCode")
     private String statusCode;
